requirements.txt
chainlit
openai
langchain
langchain-openai
langchain-community
pypdf
python-dotenv
fpdf2
beautifulsoup4
